Create helpers for account settings
